requirements.txt
pytest
python-dotenv
pywin32
traitlets
win32gui==221.6
